backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
motor==3.4.0
pymongo==4.7.3
python-dotenv==1.0.1
pyjwt==2.8.0
bcrypt==4.1.3
scikit-learn==1.5.0
numpy==1.26.4
joblib==1.4.2
httpx==0.27.0
pydantic==2.7.4
